requirements.txt
torch~=1.9.1
pandas~=1.3.3
tqdm~=4.62.3
numpy~=1.21.2
wandb~=0.12.4
scikit-learn~=1.0
PyYAML~=5.4.1
scipy~=1.7.1